renamed to module geo-compose
