+++ settings.gradle.kts
@@ -75,8 +75,8 @@ include(":bom")
 include(":composable-map")
 include(":indoors")
 include(":floor-filter-app")
-include(":geo-composables")
-include(":map-composable-app")
+include(":geo-compose")
+include(":map-compose-app")
 
 projects.forEach {
     project(":$it").projectDir = File(rootDir, "toolkit/$it")
@@ -87,5 +87,5 @@ project(":bom").projectDir = File(rootDir, "bom")
 project(":composable-map").projectDir = File(rootDir, "toolkit/composable-map")
 project(":indoors").projectDir = File(rootDir, "toolkit/indoors")
 project(":floor-filter-app").projectDir = File(rootDir, "microapps/FloorFilterApp/app")
-project(":geo-composables").projectDir = File(rootDir, "toolkit/geo-composables")
-project(":map-composable-app").projectDir = File(rootDir, "microapps/MapComposableApp/app")
+project(":geo-compose").projectDir = File(rootDir, "toolkit/geo-compose")
+project(":map-compose-app").projectDir = File(rootDir, "microapps/mapcomposeapp/app")
