+++ toolkit/geo-compose/src/main/AndroidManifest.xml
